requirements.txt
celery
Django==4.1
django-jet-reboot==1.3.3
django-admin-sortable2==2.1.4
djangorestframework==3.14.0
django-filter==22.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-celery-beat==2.5.0
gunicorn==20.1.0
requests
rabbitmq==0.2.0
psycopg2-binary==2.9.3
Pillow==9.2.0
drf-spectacular==0.25.1
python-dotenv==0.21.0
django-import-export==3.1.0
django-simple-history==3.3.0
redis
